refactor: Phase 2 — extract 7 shared helpers to FreeAgencyTableRendererView
- New FreeAgencyTableRendererInterface (7 render methods)
- New FreeAgencyTableRendererView implementing the interface
- FreeAgencyView constructor injects tableRenderer collaborator;
  all 7 private helper bodies deleted, calls rewired to tableRenderer
- RequireTrustedAnnotationRule extended to allow $this->property->method()
  chains (trusted-by-construction injected collaborator pattern)
- TrustedSafe fixture adds collaborator chain test case
- Baseline regenerated (4 stale suppress entries removed)

5 golden tests green, PHPStan clean, fixtures untouched.
