+++ ibl5/tests/PHPStanRules/Fixtures/TrustedSafe.php
@@ -2,8 +2,23 @@
 
 declare(strict_types=1);
 
+class TrustedSafeCollaborator
+{
+    public function renderCell(): string
+    {
+        return '<td></td>';
+    }
+}
+
 class TrustedSafeFixture
 {
+    private TrustedSafeCollaborator $collaborator;
+
+    public function __construct()
+    {
+        $this->collaborator = new TrustedSafeCollaborator();
+    }
+
     public function render(int $count): void
     {
         HtmlSanitizer::trusted('<b>literal html</b>');
@@ -11,6 +26,7 @@ class TrustedSafeFixture
         HtmlSanitizer::trusted((float) $count);
         HtmlSanitizer::trusted((bool) $count);
         HtmlSanitizer::trusted($this->renderRow());
+        HtmlSanitizer::trusted($this->collaborator->renderCell());
     }
 
     private function renderRow(): string
